All language files side-by-side to ease the task of ensuring they are all the same.  Realizing the constant definitions will need to be changed in the future. The .NET developeres should not work on any language except English before all development groups come to a consensus as to what the constants should be named.
There are still several issues blocking Release 1 and there will be changes to the language files as a result.

Subsequent to that release, the entire language global files are going to need to be changed to make sense as per RQuadlings's comments. If we are going to use .NET resource files after release 1, then the .NET group should not waste time on other languages.