Issues 31, 34, and 45.
Login may be set NOT to remember logins, password length or password bit entropy will be enforced.

If lg_password_min_bits > 0 then bit entropy will be used, else if lg_password_min_length > 0 and lg_password_min_bits = 0 the password length will be used.

As a result of not echoing the password to the screen and of hashing the password before entering in the database, no filtering is done on password. The only debug write of password is HTMLEncoded.

Set_new_password and change_password need to be modified.